+++ src/StorageBase/Action/Identity/Persist/IdentityPersistDeletePayload.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\Base\Action\Identity\Persist;
 
-class IdentityPersistDeletePayload extends IdentityPersistPayloadContract
+final class IdentityPersistDeletePayload extends IdentityPersistPayloadContract
 {
 }