+++ src/StorageBase/Action/PortalExtension/Deactivate/PortalExtensionDeactivatePayload.php
@@ -4,11 +4,15 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate;
 
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 
-class PortalExtensionDeactivatePayload
+final class PortalExtensionDeactivatePayload implements AttachmentAwareInterface
 {
+    use AttachmentAwareTrait;
+
     private PortalNodeKeyInterface $portalNodeKey;
 
     /**